Add redux store, redux toolkit api integration, eslint rules, error handling, loading states, infinite scroll
