+++ .env.example
@@ -1,2 +1,4 @@
-MOVIE_DB_API_URL=https://api.themoviedb.org/3/
-MOVIE_DB_API_TOKEN=<tmdb_token>
+FETCH_API_ENDPOINT=https://api.themoviedb.org/3/
+IMAGE_API_ENDPOINT=https://image.tmdb.org/t/p/
+API_TOKEN=<your_token>
+NODE_ENV=<dev|debug|prod>
