+++ __tests__/App.test.tsx
@@ -2,15 +2,15 @@
  * @format
  */
 
-import 'react-native';
-import React from 'react';
-import App from '../src/App';
-
 // Note: import explicitly to use the types shipped with jest.
-import {it} from '@jest/globals';
-
+import { it } from '@jest/globals';
+import React from 'react';
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
+
+import App from '../src/App';
+
+import 'react-native';
 
 it('renders correctly', () => {
   renderer.create(<App />);
