+++ index.js
@@ -3,7 +3,8 @@
  */
 
 import { AppRegistry } from 'react-native';
-import App from './src/App';
+
 import { name as appName } from './app.json';
+import App from './src/App';
 
 AppRegistry.registerComponent(appName, () => App);
