+++ package.json
@@ -6,32 +6,43 @@
     "android": "react-native run-android",
     "ios": "react-native run-ios",
     "lint": "eslint .",
+    "lint-fix": "eslint . --fix",
     "start": "react-native start",
     "test": "jest"
   },
   "dependencies": {
     "@react-navigation/native": "^6.1.14",
     "@react-navigation/stack": "^6.3.25",
+    "@reduxjs/toolkit": "^2.2.1",
     "react": "18.2.0",
     "react-native": "0.73.5",
+    "react-native-config": "^1.5.1",
     "react-native-gesture-handler": "^2.15.0",
     "react-native-safe-area-context": "^4.9.0",
-    "react-native-screens": "^3.29.0"
+    "react-native-screens": "^3.29.0",
+    "react-redux": "^9.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
     "@babel/preset-env": "^7.20.0",
     "@babel/runtime": "^7.20.0",
     "@react-native/babel-preset": "0.73.21",
-    "@react-native/eslint-config": "0.73.2",
     "@react-native/metro-config": "0.73.5",
     "@react-native/typescript-config": "0.73.1",
     "@types/react": "^18.2.6",
     "@types/react-test-renderer": "^18.0.0",
     "babel-jest": "^29.6.3",
     "eslint": "^8.19.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-typescript": "^18.0.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-react": "^7.34.0",
+    "eslint-plugin-simple-import-sort": "^12.0.0",
     "jest": "^29.6.3",
-    "prettier": "2.8.8",
+    "prettier": "^3.2.5",
     "react-test-renderer": "18.2.0",
     "typescript": "5.0.4"
   },
