+++ src/components/MovieList/MovieList.tsx
@@ -1,31 +1,46 @@
-import { FlatList, SafeAreaView, StyleSheet, Text, View } from 'react-native';
-import { Dimensions } from 'react-native';
+import {
+  ActivityIndicator,
+  Dimensions,
+  FlatList,
+  StyleSheet,
+  View,
+} from 'react-native';
 
-import { IMovieListItem } from '../../types/domain';
+import { IMovieListItem } from '../../types/movie';
 import MovieItem from './MovieItem';
 
 interface IMovieListProps {
   data: IMovieListItem[];
   onSelectMovie: (id: string) => void;
+  fetchMoreMovies: () => void;
+  isLoading: boolean;
 }
-const { height, width } = Dimensions.get('window');
+const { width } = Dimensions.get('window');
 const numColumns = Math.floor(width / (125 + 5 * 2));
 
-const MovieList = ({ data, onSelectMovie }: IMovieListProps) => {
-  return (
-    <View style={styles.flatListContainer}>
-      <FlatList
-        data={data}
-        numColumns={numColumns}
-        keyExtractor={(item: IMovieListItem) => `${item.id}`}
-        renderItem={({ item }) => (
-          <MovieItem onSelectMovie={onSelectMovie} {...item} />
-        )}
-        showsVerticalScrollIndicator={false}
-      />
-    </View>
-  );
-};
+const MovieList = ({
+  data,
+  onSelectMovie,
+  fetchMoreMovies,
+  isLoading,
+}: IMovieListProps) => (
+  <View style={styles.flatListContainer}>
+    <FlatList
+      data={data}
+      numColumns={numColumns}
+      keyExtractor={(item: IMovieListItem) => `${item.id}`}
+      renderItem={({ item }) => (
+        <MovieItem onSelectMovie={onSelectMovie} {...item} />
+      )}
+      showsVerticalScrollIndicator={false}
+      onEndReached={fetchMoreMovies}
+      onEndReachedThreshold={0.1}
+      ListFooterComponent={
+        isLoading ? <ActivityIndicator size="large" /> : null
+      }
+    />
+  </View>
+);
 
 const styles = StyleSheet.create({
   flatListContainer: {
