+++ src/config/index.ts
@@ -1,0 +1,8 @@
+import Config from 'react-native-config';
+
+const config = {
+  FETCH_API_ENDPOINT: Config.FETCH_API_ENDPOINT,
+  IMAGE_API_ENDPOINT: Config.IMAGE_API_ENDPOINT,
+  ACCESS_TOKEN: Config.API_TOKEN,
+};
+export default config;
