+++ src/types/movieDetails.ts
@@ -1,36 +1,4 @@
-export interface IMovieListItem {
-  vote_count?: number;
-  id?: number;
-  video?: boolean;
-  vote_average?: number;
-  title?: string;
-  popularity?: number;
-  poster_path?: string;
-  original_language?: string;
-  original_title?: string;
-  genre_ids?: number[];
-  backdrop_path?: string;
-  adult?: boolean;
-  overview?: string;
-  release_date?: string;
-}
-
-export interface ICast {
-  adult?: boolean;
-  gender?: number;
-  id?: number;
-  known_for_department?: string;
-  name?: string;
-  original_name?: string;
-  popularity?: number;
-  profile_path?: string | null;
-  cast_id?: number;
-  character?: string;
-  credit_id?: string;
-  order?: number;
-}
-
-export interface IMovieDetailsResponse {
+export interface IMovieDetails {
   adult: boolean;
   backdrop_path: string | null;
   belongs_to_collection: null | {
@@ -68,6 +36,30 @@ export interface IMovieDetailsResponse {
   vote_count: number;
 }
 
-export interface IMovieDetails extends IMovieDetailsResponse {
+export interface IFetchMovieDetailsResponse extends IMovieDetails {
+  credits: {
+    cast: ICast[];
+  };
+}
+
+export interface IMovieData extends IMovieDetails {
+  credits: {
+    cast: ICast[];
+  };
   casts: ICast[];
+}
+
+export interface ICast {
+  adult?: boolean;
+  gender?: number;
+  id?: number;
+  known_for_department?: string;
+  name?: string;
+  original_name?: string;
+  popularity?: number;
+  profile_path?: string | null;
+  cast_id?: number;
+  character?: string;
+  credit_id?: string;
+  order?: number;
 }
