+++ tsconfig.json
@@ -1,3 +1,9 @@
 {
-  "extends": "@react-native/typescript-config/tsconfig.json"
+  "extends": "@react-native/typescript-config/tsconfig.json",
+  "include": [
+    "src",
+    ".eslintrc.js",
+    "__tests__",
+    "__dummy_data__"
+  ],
 }
